+++ services/cd-service/pkg/repository/testutil/testutil.go
@@ -18,7 +18,9 @@ package testutil
 
 import (
 	"context"
+	"errors"
 	"fmt"
+	"os"
 	"strings"
 	"time"
 
@@ -127,4 +129,22 @@ func NewIncrementalUUIDGenerator() uuid.GenerateUUIDs {
 		gen: &fakeGenBase,
 	}
 	return fakeGen
+}
+
+// CreateMigrationsPath detects if it's running withing earthly/CI or locally and adapts the path to the migrations accordingly
+func CreateMigrationsPath() (string, error) {
+	const subDir = "/database/migrations/sqlite"
+	_, err := os.Stat("/kp")
+	if err != nil {
+		if errors.Is(err, os.ErrNotExist) {
+			wd, err := os.Getwd()
+			if err != nil {
+				return "", err
+			}
+			// this ".." sequence is necessary, because Getwd() returns the path of this go file (when running in an idea like goland):
+			return wd + "/../../../.." + subDir, nil
+		}
+		return "", err
+	}
+	return "/kp" + subDir, nil
 }
