+++ services/cd-service/pkg/service/git_test.go
@@ -732,9 +732,14 @@ func TestGetCommitInfo(t *testing.T) {
 		tc := tc
 		t.Run(tc.name, func(t *testing.T) {
 			shutdown := make(chan struct{}, 1)
+			dir, err := testutil.CreateMigrationsPath()
+			if err != nil {
+				t.Fatalf("setup error could not detect dir \n%v", err)
+				return
+			}
 			cfg := db.DBConfig{
 				DriverName:     "sqlite3",
-				MigrationsPath: "/kp/database/migrations",
+				MigrationsPath: dir,
 			}
 			repo, err := setupRepositoryTestWithDB(t, &cfg)
